mspar specific code is a bit more decoupled from ms code now
